services/chat-service/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
httpx>=0.27.0
python-dotenv>=1.0.1
psycopg[binary]>=3.2.0
